Сделал суперкласс Ship. Поломалось движение за курсором.
