--- core/src/ru/geekbrains/sprites/BulletPool.java
@@ -1,16 +1,0 @@
-package ru.geekbrains.sprites;
-
-import ru.geekbrains.base.SpritesPool;
-
-public class BulletPool extends SpritesPool<Bullet> {
-
-    @Override
-    protected Bullet newObject() {
-        return new Bullet();
-    }
-
-    @Override
-    protected void log() {
-        System.out.println("Bullet active/free: " + activeObjects.size() + "/" + freeObjects.size());
-    }
-}
